Generate PRPS for RUV-III

This function is a part of the data analysis functionality of tgcapkg. It creates pseudo-replicates of pseudo-samples (PRPS) for unwanted variations like Library size, Batches and Purity in TCGA Pan Cancer Datasets with Cancer biology like Breast Cancer data (BRCA), Lung Cancer (LUAD), Colon Cancer (COAD) and Rectum Cancer (READ). In the function batch refers to the source of Batch Effect variation like Time and Plate which captures variation across biology while factors like Purity captures variation within biology.

createPRPS(
  data,
  librarySize = "ls",
  batch,
  biology = "Subtypes",
  purity,
  include.ls,
  include.purity,
  minSamplesPerBatchPS,
  minSamplesForPuirtyPS,
  minSamplesForPurityPerBiology,
  minSamplesForLibrarySizePerBatch,
  minSamplesForLibrarySizePS
)

Arguments

data
S4 data object

librarySize
character: Library Size variable in input data object. Default is 'ls'.

batch
character: Batch effect factors. In current package version batch can take values like 'Year', 'Plate' or both

biology
character: Biology of cancer type. TCGA datasets have biology for only four Cancer types i.e. Lung (LUAD), Breast (BRCA), Rectum (READ) & Colon (COAD). So the function supports only these four datasets for RUV-III and PRPS analysis. Default is 'Subtypes'.

purity
character: Purity variable in input data object

include.ls
logical: Do we need to consider library size in creating pseudo samples

include.purity
logical: Do we need to consider purity in creating pseudo samples

minSamplesPerBatchPS
numeric: Minimum number of samples per batch for creating Pseudo Samples

minSamplesForPuirtyPS
numeric: Minimum number of samples for creating Pseudo Samples for purity.

minSamplesForPurityPerBiology
numeric: Number of samples for purity per biology for creating Pseudo Samples

minSamplesForLibrarySizePerBatch
numeric: Number of samples for library size per batch for creating Pseudo Samples

minSamplesForLibrarySizePS
numeric: Minimum number of samples for creating Pseudo Samples for library size

Value

A S4 list object with the Pseudo replicate for pseudo samples for different batches, library size and purity.

Examples

if (FALSE) {
createPRPS(data=brca.data, batch=c('Year', 'Plates'), purity='Purity_singscore',include.ls=TRUE, include.purity=TRUE,
minSamplesPerBatchPS = 3, minSamplesForPuirtyPS = 3, minSamplesForPurityPerBiology = 12,
minSamplesForLibrarySizePerBatch = 6,minSamplesForLibrarySizePS = 3)
}