Generate RUV-III PRPS data

This function is a part of the data analysis functionality of tgcapkg. It creates pseudo samples for on Library size, Batches and Purity.

get.prps(
  data,
  batch,
  biology,
  purity,
  include.ls,
  include.purity,
  n.ls.ps,
  n.sample.ls,
  n.sample.batch,
  n.sample.purity
)

Arguments

data
S4 data object

batch
character: Batch effect factors. In current package version batch can be given values like 'Year', 'Plate' or both.

biology
character: Biology of cancer type. In current package version biology for cancer type Breast, Rectum & Colon is considered.

purity
character: Purity column name

include.ls
logical: Do we need to consider library size in creating pseudo samples.

include.purity
logical: Do we need to consider purity in creating pseudo samples.

n.ls.ps
numeric: Number of samples for library size considered within plate.

n.sample.ls
numeric: Minimum number of samples within library size pseudo sample size.

n.sample.batch
numeric: Number of samples per batch.

n.sample.purity
numeric: Number of samples for purity.

Value

A S4 list object with the Pseudo replicate for pseudo samples for different batches.

Examples

if (FALSE) {
get.prps(data=brca.data, batch=c('Year', 'Plates'), biology='biology', purity=NULL, include.ls=T, include.purity=F, n.ls.ps=10, n.sample.batch=3, n.sample.purity=0, n.sample.ls=3)
get.prps(data=brca.data, batch=c('Year', 'Plates'), biology='biology', purity='Purity_singscore', include.ls=T, include.purity=T, n.ls.ps=10, n.sample.batch=3, n.sample.purity=3, n.sample.ls=3)
}