Correlation Analysis between PCs and variation types

This function is a part of the data analysis functionality of tgcapkg. It helps user to run regression between bias in TCGA RNA-seq like librarysize and purity with PCs from get.pca. The output is a linear plot that compares the three assays in SummarizedExperiment TGCA Cancer data across n PCs and R-sq. It also runs vector correlation stats::cancor between time and n PCs with the same linear output explaining variation explained by different variation types.

pca.corr(pca.data, data, type, nPCs)

Arguments

pca.data
list: PCA output from get.pca.

data
S4 data object

type
character: The response variable to lm model. groups included are 'librarysize', 'purity' and 'time'.

nPCs
numeric: Number of PCs that needs to be used for regression

Value

Linear Plot the compares the correlation between library size (or Purity, time) and PCs across three datasets. When output is stored in a object the user can also access values used to plot the linear graphs.

Examples

if (FALSE) {
pca.corr(pca.data, data = brca.data, type = "purity", nPCs = 10)
df <- pca.corr(pca.data, data = brca.data, type = "time", nPCs = 8)
df
}